Change structure of the exported attachments tarball: To account for the case that an item can have two attachments with the same filename. In that case only one of them (the last one) would be exported (as the other ones would be overwritten when exported sequentially).

+++ bw-restore-attachments.sh
@@ -66,10 +66,13 @@ while [ "${1}" != "" ] ; do
                 if [ "${DESTINATION_ITEM_ID_WITH_ATTACHMENTS_ALREADY}" != "" ] ; then
                     echo2 "WARNING: item \`${DESTINATION_ITEM_ID}/${EXPORTED_ITEM_FIELDS}' already has attachments."
                 else
-                    for ATTACHMENT_FILE in * ; do
-                        echoprompt "bw create attachment --file '${ATTACHMENT_FILE}' --itemid ${DESTINATION_ITEM_ID}"
-                        bw create attachment --file "${ATTACHMENT_FILE}" --itemid ${DESTINATION_ITEM_ID}
-                        rm -vf "${ATTACHMENT_FILE}"
+                    for ATTACHMENT_FILE in */* * ; do
+                        if [ -f "${ATTACHMENT_FILE}" ] ; then
+                            echoprompt "bw create attachment --file '${ATTACHMENT_FILE}' --itemid ${DESTINATION_ITEM_ID}"
+                            bw create attachment --file "${ATTACHMENT_FILE}" --itemid ${DESTINATION_ITEM_ID}
+                        fi
+                        test -f "${ATTACHMENT_FILE}" && rm -vf "${ATTACHMENT_FILE}"
+                        test -d "${ATTACHMENT_FILE}" && rmdir -v --ignore-fail-on-non-empty "${ATTACHMENT_FILE}"
                     done
                 fi
             fi
